+++ crates/rspack_plugin_mf/src/container/remote_runtime_module.rs
@@ -1,8 +1,7 @@
 use rspack_core::{
   impl_runtime_module,
   rspack_sources::{BoxSource, RawSource, SourceExt},
-  ChunkUkey, Compilation, DependenciesBlock, RuntimeGlobals, RuntimeModule, RuntimeModuleStage,
-  SourceType,
+  ChunkUkey, Compilation, DependenciesBlock, RuntimeModule, RuntimeModuleStage, SourceType,
 };
 use rspack_identifier::{Identifiable, Identifier};
 use rustc_hash::FxHashMap;
@@ -85,13 +84,12 @@ impl RuntimeModule for RemoteRuntimeModule {
     }
     RawSource::from(format!(
       r#"
-__webpack_require__.MF.remotesLoadingData = {{ chunkMapping: {chunk_mapping}, moduleIdToRemoteDataMapping: {id_to_external_and_name_mapping} }};
-{ensure_chunk_handlers}.remotes = function(chunkId, promises) {{ return {remotes_fn}(chunkId, promises); }};
+__webpack_require__.remotesLoadingData = {{ chunkMapping: {chunk_mapping}, moduleIdToRemoteDataMapping: {id_to_external_and_name_mapping} }};
+{remotes_loading_impl}
 "#,
       chunk_mapping = json_stringify(&chunk_to_remotes_mapping),
       id_to_external_and_name_mapping = json_stringify(&id_to_external_and_name_mapping),
-      ensure_chunk_handlers = RuntimeGlobals::ENSURE_CHUNK_HANDLERS,
-      remotes_fn = "__webpack_require__.MF.remotesLoading",
+      remotes_loading_impl = include_str!("./remotesLoading.js"),
     ))
     .boxed()
   }
