+++ packages/rspack/src/container/ContainerPlugin.ts
@@ -14,7 +14,6 @@ import {
 import { isNil } from "../util";
 import { parseOptions } from "../container/options";
 import { Compiler } from "../Compiler";
-import { ModuleFederationRuntimePlugin } from "./ModuleFederationRuntimePlugin";
 
 export type ContainerPluginOptions = {
 	exposes: Exposes;
@@ -72,10 +71,6 @@ export class ContainerPlugin extends RspackBuiltinPlugin {
 		if (!compiler.options.output.enabledLibraryTypes!.includes(library.type)) {
 			compiler.options.output.enabledLibraryTypes!.push(library.type);
 		}
-		ModuleFederationRuntimePlugin.addPlugin(
-			compiler,
-			require.resolve("../sharing/initializeSharing.js")
-		);
 		return createBuiltinPlugin(this.name, this._options);
 	}
 }
