+++ packages/rspack/src/rspackOptionsApply.ts
@@ -38,7 +38,6 @@ import {
 	SplitChunksPlugin,
 	OldSplitChunksPlugin
 } from "./builtin-plugin";
-import { ModuleFederationRuntimePlugin } from "./container/ModuleFederationRuntimePlugin";
 import { WorkerPlugin } from "./builtin-plugin/WorkerPlugin";
 
 export function optionsApply_compat(
@@ -174,8 +173,6 @@ export class RspackOptionsApply {
 			"options.context should have value after `applyRspackOptionsDefaults`"
 		);
 		compiler.hooks.entryOption.call(options.context, options.entry);
-
-		new ModuleFederationRuntimePlugin().apply(compiler);
 
 		const { minimize, minimizer } = options.optimization;
 		if (minimize && minimizer) {
