+++ packages/rspack/src/sharing/ProvideSharedPlugin.ts
@@ -6,7 +6,6 @@ import {
 } from "../builtin-plugin/base";
 import { parseOptions } from "../container/options";
 import { Compiler } from "../Compiler";
-import { ModuleFederationRuntimePlugin } from "../container/ModuleFederationRuntimePlugin";
 
 export type ProvideSharedPluginOptions = {
 	provides: Provides;
@@ -53,10 +52,6 @@ export class ProvideSharedPlugin extends RspackBuiltinPlugin {
 	}
 
 	raw(compiler: Compiler): BuiltinPlugin {
-		ModuleFederationRuntimePlugin.addPlugin(
-			compiler,
-			require.resolve("./initializeSharing.js")
-		);
 		return createBuiltinPlugin(this.name, this._options);
 	}
 }
